Makefile: Match board name with compatible string and install sysfs_attr helper
There is an script that parses /proc/device-tree/compatible and if
matches with the name of the board script runs the test suite. The
Samsung Chromebook Plus reports the following compatibles:

 google,kevin-rev15 google,kevin-rev14 google,kevin-rev13
 google,kevin-rev12 google,kevin-rev11 google,kevin-rev10
 google,kevin-rev9 google,kevin-rev8 google,kevin-rev7
 google,kevin-rev6 google,kevin google,gru rockchip,rk3399

Being google,kevin the more generict this patch renames the board test
script to this name.

The patch also adds the assert_sysfs_attr_present helper script to the
Makefile install target because it is used by the google,kevin test
suite.

Fixes: 1c4202515231 ("Makefile: fixes board name for google,kevin")
Fixes: c5d93c41f893 ("boards: Add Samsung Chromebook Plus test script")
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>