helpers: device_present: Report driverless devices as blocked
Typical use of assert_device_present is to first assert_driver_present
and then check for the individual devices. Let the driver check fail and
report any of the device tests as blocked if the driver is not present.

This makes it clearer in the output that a device test failed due to the
lack of the driver, rather than some other aspect.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>