helpers/bootrr: Silence the lava-test-case test
`command -v` is invoked to check if lava-test-case exists, to know if we
need to fake it using `echo`. But in the event that lava-test-case
actually exist it also prints the full path of the executable.

Pipe the output to /dev/null, as we only care about the exit code.

Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>