Makefile: Override owner of files in cpio archive
In the typical scenario where the bootrr cpio archive is concatenated
onto some ramdisk, attributes of this cpio archive will overwrite those
previously extracted.

When used in combination with a ramdisk with systemd, this results in
many services failing to start, because the namespaced services does not
have permission to open e.g. /

Fix this my overriding the owner and group.

Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>